+++ VentasProductos/src/ventasproductos/FamiliaProducto.java
@@ -1,0 +1,12 @@
+package ventasproductos;
+
+/**
+ *
+ * @author adrian
+ */
+public enum FamiliaProducto {
+    LICOR, 
+    VEGETALES, 
+    FRUTAS,
+    TECNOLOGIA
+}