+++ VentasProductos/src/ventasproductos/VentasProductos.java
@@ -1,0 +1,14 @@
+
+package ventasproductos;
+
+/**
+ *
+ * @author Adrian Verdugo
+ */
+public class VentasProductos {
+
+    public static void main(String[] args) {
+        new Principal().setVisible(true);
+    }
+    
+}